+++ STUDIO/branches/redesign/org.openl.rules.webstudio/src/org/openl/rules/ui/ProjectModel.java
@@ -173,7 +173,7 @@ public class ProjectModel {
                         }
 
                         @Override
-                        public void runNtimes(int times) throws Exception {
+                        public void runNtimes(long times) throws Exception {
                             testSuite.invoke(target, env, times);
                         }
 
@@ -216,7 +216,7 @@ public class ProjectModel {
             }
 
             @Override
-            public void runNtimes(int times) throws Exception {
+            public void runNtimes(long times) throws Exception {
                 try {
                     testSuite.getTest(testIndex).runTest(target, env, times);
                 } catch (Throwable t) {
@@ -271,7 +271,7 @@ public class ProjectModel {
                         }
 
                         @Override
-                        public void runNtimes(int times) throws Exception {
+                        public void runNtimes(long times) throws Exception {
                             bm.invokeBenchmark(target, params, env, times);
                         }
